requirements.txt
slack-bolt>=1.18.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pytz>=2024.1
